underWater/internal api/requirements.txt
Django==4.2.13
requests>=2.31.0
python-dotenv>=1.0.0
SQLAlchemy>=2.0.25
psycopg2-binary>=2.9.9
alembic>=1.13.1
gunicorn>=22.0.0
PyJWT>=2.8.0
